requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pygithub==2.1.1
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
playwright==1.40.0
aiofiles==23.2.1
asyncio==3.4.3